Extract task builder from ITask for readability
